check

+++ hxline/services/PeopleService.go
@@ -5,28 +5,32 @@ import (
 	repo "hxline/repositories"
 )
 
-func Create(people model.Person) {
-	repo.Insert(people)
+func Create(people model.Person) (err error) {
+	err = repo.Insert(people)
+
+	return
 }
 
-func Update(people model.Person) {
-	repo.Update(people)
-}
-
-func Delete(id string) {
-	repo.Delete(id)
-}
-
-func ReadAllPeople() (peoples []model.Person) {
-	peopleRows := repo.GetAllPeople()
-
-	for peopleRows.Next() {
-		var id string
-		var name string
-		var age int
-		peopleRows.Scan(&id, &name, &age)
-		peoples = append(peoples, model.Person{ID: id, Name: name, Age: age})
+func Update(people model.Person) (result []model.Person, err error) {
+	result, err = repo.GetPerson(people.ID)
+	if err == nil && len(result) > 0 {
+		err = repo.Update(people)
 	}
+
+	return
+}
+
+func Delete(id string) (result []model.Person, err error) {
+	result, err = repo.GetPerson(id)
+	if err == nil && len(result) > 0 {
+		err = repo.Delete(id)
+	}
+
+	return
+}
+
+func ReadAllPeople() (peoples []model.Person, err error) {
+	peoples, err = repo.GetAllPeople()
 
 	return
 }
